main created
A main file was created that will be able to use the scramble grabbing features and a display cube file was made to show the output of the cube. This will call some of the movement functions so that it can make the turns by itself and show what happens each time.